Add Code: Does Robot walk in Circle?
The function checks if the robot has returned to the starting position and/or if it's facing north. If either condition is met, it is determined that the robot's path forms a circle, and "YES" is added to the result list. Otherwise, "NO" is added. The function returns this list of results corresponding to each command sequence.